Revert "Extra Content: extra content to be removed"
This reverts commit d01c643eadb2afb532558ccb2fbef6b471fc21aa.
