--- examples/stable-diffusion/quantization/measure_config.json
@@ -1,6 +1,0 @@
-{
-    "method": "HOOKS",
-    "mode": "MEASURE",
-    "observer": "maxabs",
-    "dump_stats_path": "./quantization/measure/fp8"
-}
