--- examples/stable-diffusion/quantization/quant_config.json
@@ -1,7 +1,0 @@
-{
-    "method": "HOOKS",
-    "mode": "QUANTIZE",
-    "observer": "maxabs",
-    "scale_method": "maxabs_hw",
-    "dump_stats_path": "./quantization/measure/fp8"
-}
