+++ examples/stable-diffusion/training/textual_inversion_sdxl.py
