+++ examples/stable-diffusion/training/train_controlnet.py
