+++ examples/stable-diffusion/training/train_dreambooth_lora_sdxl.py
