+++ examples/stable-diffusion/unconditional_image_generation.py
