+++ examples/summarization/ds_flan_t5_z3_config_bf16.json
@@ -32,7 +32,7 @@
     "stage3_param_persistence_threshold": "auto",
     "stage3_max_live_parameters": 1e9,
     "stage3_max_reuse_distance": 1e9,
-    "stage3_gather_16bit_weights_on_model_save": true
+    "stage3_gather_16bit_weights_on_model_save": false
   },
   "gradient_accumulation_steps": "auto",
   "gradient_clipping": "auto",
