--- examples/text-generation/quantization_config/maxabs_quant_const_scales.json
@@ -1,8 +1,0 @@
-{
-    "method": "HOOKS",
-    "mode": "QUANTIZE",
-    "observer": "maxabs",
-    "scale_method": "maxabs_hw",
-    "dump_stats_path": "./hqt_output/measure",
-    "scale_format": "const"
-}
