--- examples/text-generation/quantization_config/maxabs_quant_dynamic_quantization.json
@@ -1,13 +1,0 @@
-{
-    "mode": "QUANTIZE",
-    "scale_method": "ACT_MAXABS_PCS_POW2_WEIGHT_MAXABS_PTS_POW2_HW",
-    "scale_format": "CONST",
-    "allowlist": {
-        "types": [],
-        "names": [
-            "mlp"
-        ]
-    },
-    "dynamic_quantization": "True",
-    "dump_stats_path": "./hqt_output/measure"
-}
