--- examples/text-generation/quantization_config/maxabs_quant_dynamic_quantization_pts.json
@@ -1,13 +1,0 @@
-{
-    "mode": "QUANTIZE",
-    "scale_method": "maxabs_pow2",
-    "scale_format": "CONST",
-    "allowlist": {
-        "types": [],
-        "names": [
-            "mlp"
-        ]
-    },
-    "dynamic_quantization": "True",
-    "dump_stats_path": "./hqt_output/measure"
-}
