--- examples/text-generation/quantization_config/maxabs_quant_qdq.json
@@ -1,9 +1,0 @@
-{
-    "method": "HOOKS",
-    "mode": "QUANTIZE",
-    "observer": "maxabs",
-    "scale_method": "maxabs_hw",
-    "scale_format": "SCALAR",
-    "dump_stats_path": "./hqt_output/measure",
-    "use_qdq": "True"
-}
