+++ optimum/habana/accelerate/utils/__init__.py
@@ -1,6 +1,1 @@
-from .other import extract_model_from_parallel
-from .transformer_engine import (
-    FP8ContextWrapper,
-    convert_model,
-    get_fp8_recipe,
-)
+from .transformer_engine import FP8ContextWrapper, convert_model, get_fp8_recipe
