+++ optimum/habana/transformers/generation/configuration_utils.py
@@ -39,8 +39,6 @@ class GaudiGenerationConfig(GenerationConfig):
         Whether to enable causal_mask if use Habana flash attention.
     flash_attention_fast_softmax_mode (`bool`, *optional*):
         Whether to use fast softmax with reduced precision if use Habana flash attention.
-    use_mark_dynamic (`bool`, *optional*):
-        Mark required tensor(s) as dynamic.
     attn_batch_split (`int`, *optional*):
         Specify the batch size split for attention and mlp layers. 1 for no split. This is enabled only for prompt.
     logits_bf16 (`bool`, *optional*):
@@ -65,6 +63,5 @@ class GaudiGenerationConfig(GenerationConfig):
         self.flash_attention_fast_softmax = kwargs.get("flash_attention_fast_softmax", None)
         self.use_fused_rope = kwargs.get("use_fused_rope", None)
         self.valid_sequence_lengths = kwargs.get("valid_sequence_lengths", None)
-        self.use_mark_dynamic = kwargs.get("use_mark_dynamic", None)
         self.attn_batch_split = kwargs.get("attn_batch_split", 1)
         self.logits_bf16 = kwargs.get("logits_bf16", None)
