+++ optimum/habana/transformers/models/__init__.py
@@ -213,7 +213,6 @@ from .modeling_all_models import (
     Matmul,
     apply_customized_rope_module,
     gaudi_check_and_enable_sdpa,
-    gaudi_check_support_param_buffer_assignment,
     gaudi_conv1d_forward,
     gaudi_get_extended_attention_mask,
     gaudi_invert_attention_mask,
