+++ optimum/habana/transformers/models/video_llava/modeling_video_llava.py
@@ -198,6 +198,7 @@ class GaudiVideoLlavaForConditionalGeneration(VideoLlavaForConditionalGeneration
             output_hidden_states if output_hidden_states is not None else self.config.output_hidden_states
         )
         return_dict = return_dict if return_dict is not None else self.config.use_return_dict
+
         outputs = self.language_model(
             attention_mask=attention_mask,
             position_ids=position_ids,
